enhance(?): correction in guide from concentrations with conserved aggregation
This is not directly comparable to the flux-based counterpart because of
the usage of Delta on the guide instead of a Normal distribution but the conserved
aggregation can be translated without changes.

The model implements a Normal(0, sigma) where sigma is a learned random
variable, scalar. A vector over the conditions migth be considered
after evaluating the approach.